refactor: 디렉토리 구조 변경 (#69)
`article` 디렉토리 밑에 `item`을 뒀었지만 도메인 구조의 변경으로 인해 더 이상 `article`에만 속하는게 아니기에 따로 두기로 함
